compressor: Move aux functions to its own module
